fixed add button errors, delete button errors not solved yet

+++ app/src/main/java/com/example/user/sqlitedemo/UsersDataSource.java
@@ -5,6 +5,7 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
+import android.util.Log;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,6 +20,8 @@ import java.util.List;
  * the persistence layer.
  */
 class UsersDataSource {
+
+    private static final String TAG = "**Data Source Class**";
 
     private SQLiteDatabase database;
     private DatabaseHelper helper;
@@ -38,23 +41,36 @@ class UsersDataSource {
         helper.close();
     }
 
-     User createUser(String username/*, String password*/){
+     User createUser(String username/*, String password*/){//shudhu username input nicchi for trial
         //User user = new User(email, password);//redundant
         ContentValues values = new ContentValues();
         values.put("username", username);
         //values.put("password", password);
-        long insertId = database.insert("Users", null, values);//ERROR//
-        Cursor cursor = database.query("Users", allColumns, "_id  = " + insertId, null, null, null, null);
-        cursor.moveToFirst();
-        User user = cursorToUser(cursor);//ERROR//
-        cursor.close();
+        long id = database.insert("Users", null, values);//ERROR//
+         //^ arguments : tableName, nullColumnHack, ContentValues
+         //returns the row id of the newly created row
+         Cursor cursor = null;
+         User user = null;
+         if(id != -1) {
+             cursor = database.query("Users", allColumns, "_id  = " + id, null, null, null, null);
+             cursor.moveToFirst();
+             user = cursorToUser(cursor);//ERROR//
+             cursor.close();
+         }else{
+             Log.e(TAG, "id returned -1");
+         }
         return user;
     }
 
      void deleteUser(User user){
-        long id = user.getId();
-        database.delete("Users", "_id ="+id, null);//delete the given id (matching with the "_id" column) from the Users table
-    }
+        if (user != null) {
+            long id = user.getId();
+            Log.e(TAG, "user's id = "+String.valueOf(id));
+            database.delete("Users", "_id =" + id, null);//delete the given id (matching with the "_id" column) from the Users table
+        }else{
+            Log.e(TAG, "user argument is null");
+        }
+     }
 
      List<User> getAllUsers(){
         List<User> users = new ArrayList<User>();
